Add page template
Added the HTML, CSS, and JS needed for the template of the application.
The template includes the main parts of the application.
These range from the boards, the component used to place ships, etc.
It is important to note that this commit does NOT integrate any game logic into the UI.