Improve PC intelligence
Improved PC intelligence.
Changed the algorithm of the PC to prioritize covering more rows.
Previously, the algorithm targeted every possible placement of the largest ship not destroyed.
Now, the algorithm will target only the horizontal placements before moving onto vertical ones.
This allows the PC to cover more rows in a game.
Note, the moves performed by the PC are still not, in any way, "optimal".